+++ src/components/player-statistics-card/components/statistics-jdc-practice/components/jdc-shirt/jdc-shirt.component.js
@@ -1,0 +1,8 @@
+module.exports = {
+    onCreate(input) {
+        this.state = {
+            shirt: input.color,
+            size: input.size
+        }
+    }
+}