+++ src/pages/leg-result/components/leg-result/leg-result.component.js
@@ -1,14 +1,14 @@
-var types = require("../../../../components/scorecard/components/match_types");
+const types = require("../../../../components/scorecard/components/match_types");
 
-var moment = require("moment");
-var axios = require("axios");
-var alertify = require("../../../../util/alertify");
-var _ = require('underscore');
-var Chart = require("chart.js");
+const moment = require("moment");
+const axios = require("axios");
+const alertify = require("../../../../util/alertify");
+const _ = require('underscore');
+const Chart = require("chart.js");
 
 module.exports = {
     onCreate(input) {
-        var leg = input.leg;
+        const leg = input.leg;
 
         leg.started = moment(leg.created_at).format('YYYY-MM-DD HH:mm:ss');
         leg.finished = leg.end_time === null ? '-' : moment(leg.end_time).format('YYYY-MM-DD HH:mm:ss');
@@ -16,30 +16,27 @@ module.exports = {
         if (leg.visits.length > 0) {
             leg.duration = moment.duration(moment(leg.end_time).diff(leg.visits[0].created_at)).asMinutes().toFixed();
         }
-        for (var i = 0; i < leg.visits.length; i++) {
-            var visit = leg.visits[i];
+        for (let i = 0; i < leg.visits.length; i++) {
+            const visit = leg.visits[i];
 
-            var count = leg.players.length;
-            var idx = leg.players.indexOf(visit.player_id);
+            const count = leg.players.length;
+            const idx = leg.players.indexOf(visit.player_id);
 
-            var order = [ leg.players[idx] ];
-            for (var j = 1; j < leg.players.length; j++) {
+            const order = [ leg.players[idx] ];
+            for (let j = 1; j < leg.players.length; j++) {
                 order[j] = leg.players[(idx + j) < count ? idx + j : (j - (count - idx))];
             }
-            var scores = visit.scores[order[0]];
-            for (var j = 1; j < order.length; j++) {
+            let scores = visit.scores[order[0]];
+            for (let j = 1; j < order.length; j++) {
                 var id = order[j];
-                if (leg.is_finished) {
-                    scores += " : " + visit.scores[id];
-                } else {
-                    scores += " : " + (visit.scores[id] ? visit.scores[id] : 0);
-                }
+                scores += ` : ${visit.scores[id] ? visit.scores[id] : 0}`;
             }
             visit.score_str = scores;
         }
         this.state = {
             leg: leg,
             matchId: leg.match_id,
+            matchType: input.leg.leg_type.id || input.match.match_type.id,
             legPlayers: input.leg_players,
             players: input.players
         }
@@ -53,14 +50,14 @@ module.exports = {
         var match = this.input.match;
         var visits = leg.visits;
 
-        var chartMaxValue = match.match_type.id == types.X01 || match.match_type.id == types.X01HANDICAP ? leg.starting_score : 0;
+        var chartMaxValue = this.state.matchType === types.X01 || this.state.matchType === types.X01HANDICAP ? leg.starting_score : 0;
         var labels = [];
         var values = { }
-        for (var i = 0; i < legPlayers.length; i++) {
-            var player = legPlayers[i];
+        for (let i = 0; i < legPlayers.length; i++) {
+            const player = legPlayers[i];
 
-            var score = leg.starting_score + player.handicap
-            if (match.match_type.id == types.DARTS_AT_X) {
+            let score = leg.starting_score + player.handicap
+            if (this.state.matchType === types.DARTS_AT_X || this.state.matchType === types.GOTCHA) {
                 score = 0;
             }
             values[player.player_id] = [ score ];
@@ -68,29 +65,36 @@ module.exports = {
                 chartMaxValue = score;
             }
         }
-        var round = 0;
+        let round = 0;
         labels.push(round);
 
-        for (var i = 0; i < visits.length; i++) {
-            var visit = visits[i];
+        for (let i = 0; i < visits.length; i++) {
+            const visit = visits[i];
             if (i % legPlayers.length === 0) {
                 round++;
                 labels.push(round);
             }
-            var current = values[visit.player_id][values[visit.player_id].length - 1];
+            let current = values[visit.player_id][values[visit.player_id].length - 1];
             if (visit.is_bust) {
                 values[visit.player_id].push(current);
             }
             else {
-                if (match.match_type.id == types.SHOOTOUT || match.match_type.id == types.DARTS_AT_X || match.match_type.id === types.CRICKET ||
-                    match.match_type.id == types.AROUND_THE_CLOCK || match.match_type.id == types.AROUND_THE_WORLD || match.match_type.id == types.SHANGHAI ||
-                    match.match_type.id == types.TIC_TAC_TOE || match.match_type.id == types.BERMUDA_TRIANGLE) {
+                if (this.state.matchType === types.SHOOTOUT || this.state.matchType === types.DARTS_AT_X || this.state.matchType === types.CRICKET ||
+                    this.state.matchType === types.AROUND_THE_CLOCK || this.state.matchType === types.AROUND_THE_WORLD || this.state.matchType === types.SHANGHAI ||
+                    this.state.matchType === types.TIC_TAC_TOE || this.state.matchType === types.BERMUDA_TRIANGLE || this.state.matchType === types.JDC_PRACTICE) {
                     current = current + visit.score;
                     values[visit.player_id].push(current);
                     if (current > chartMaxValue) {
                         chartMaxValue = current;
                     }
-                } else if (match.match_type.id == types.KILL_BULL) {
+                } else if (this.state.matchType === types.GOTCHA) {
+                    if (visit.score === 0) {
+                        current = 0
+                    } else {
+                        current += visit.score;
+                    }
+                    values[visit.player_id].push(current);
+                } else if (this.state.matchType === types.KILL_BULL) {
                     if (visit.score === 0) {
                         current = leg.starting_score;
                     } else {